+++ output/changelog/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-- Generated at: 2026-01-04T14:44:11.879678
-- Total time: 55.18 seconds
+- Generated at: 2026-01-04T19:15:05.406952
+- Total time: 46.42 seconds
 - Features: 23
 - Models: 15